feat:fluxo inicial do backend com lógicas de acesso, produto e carrinho

+++ dolcelou/src/main/java/com/dolcelou/model/ProdutoEntity.java
@@ -38,5 +38,5 @@ public class ProdutoEntity {
     String descricao;
 
     @Column(nullable = false)
-    Integer quantidade;
+    double quantidade;
 }
